Change DWORD type definition from unsigned long to unsigned int - Fixes scroll bug in hud in 64 bit exe.

+++ src/K6502.h
@@ -11,7 +11,7 @@
 
 // Type definition
 #ifndef DWORD
-typedef unsigned long  DWORD;
+typedef unsigned int  DWORD;
 #endif
 
 #ifndef WORD
